Renamed wagtail.contrib.wagtailfrontendcache to wagtail.contrib.frontend_cache
