+++ wagtail/api/v2/apps.py
@@ -13,8 +13,8 @@ class WagtailAPIV2AppConfig(AppConfig):
     def ready(self):
         # Install cache purging signal handlers
         if getattr(settings, 'WAGTAILAPI_USE_FRONTENDCACHE', False):
-            if apps.is_installed('wagtail.contrib.wagtailfrontendcache'):
+            if apps.is_installed('wagtail.contrib.frontend_cache'):
                 from wagtail.api.v2.signal_handlers import register_signal_handlers
                 register_signal_handlers()
             else:
-                raise ImproperlyConfigured("The setting 'WAGTAILAPI_USE_FRONTENDCACHE' is True but 'wagtail.contrib.wagtailfrontendcache' is not in INSTALLED_APPS.")
+                raise ImproperlyConfigured("The setting 'WAGTAILAPI_USE_FRONTENDCACHE' is True but 'wagtail.contrib.frontend_cache' is not in INSTALLED_APPS.")
