+++ wagtail/api/v2/signal_handlers.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, unicode_literals
 from django.db.models.signals import post_delete, post_save
 from django.urls import reverse
 
-from wagtail.contrib.wagtailfrontendcache.utils import purge_url_from_cache
+from wagtail.contrib.frontend_cache.utils import purge_url_from_cache
 from wagtail.wagtailcore.models import get_page_models
 from wagtail.wagtailcore.signals import page_published, page_unpublished
 from wagtail.wagtaildocs.models import get_document_model
