+++ wagtail/contrib/frontend_cache/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'wagtail.contrib.frontend_cache.apps.WagtailFrontendCacheAppConfig'
