+++ wagtail/contrib/frontend_cache/signal_handlers.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, unicode_literals
 
 from django.apps import apps
 
-from wagtail.contrib.wagtailfrontendcache.utils import purge_page_from_cache
+from wagtail.contrib.frontend_cache.utils import purge_page_from_cache
 from wagtail.wagtailcore.signals import page_published, page_unpublished
