--- wagtail/contrib/wagtailfrontendcache/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'wagtail.contrib.wagtailfrontendcache.apps.WagtailFrontendCacheAppConfig'
